requirements.txt
Django
python-decouple # using for environments variables 
pillow
drf_yasg
djangorestframework
django-otp
psycopg2-binary
djangorestframework-simplejwt